requirements.txt
asgiref==3.7.2
Django==4.2.4
django-dotenv==1.4.2
django-phonenumber-field==7.1.0
phonenumberslite==8.13.18
psycopg2==2.9.6
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
